requirements.txt
streamlit>=1.35
pillow>=10.0
rembg[cpu]>=2.0.60
